Extract shell integration helpers into amux-core (#42 step 3)
Move default_shell(), inject_shell_integration(), ensure_shell_integration_dir(),
and ensure_claude_wrapper_dir() to amux-core::shell. These are pure
filesystem/env operations with no terminal dependency.

spawn_surface() stays in amux-app (returns PaneSurface which depends on
TerminalPane), but now calls shell:: helpers from amux-core.
