+++ crates/amux-core/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod config;
 pub mod keys;
 pub mod model;
+pub mod shell;
